Меняю вызов функции вывода фотографий
